requirements.txt
black==22.1.0
click==8.0.4
colorama==0.4.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
tomli==2.0.1
